pyproject.toml
[project]
name = "pharma_back"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.119.0,<0.120.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "bcrypt (==4.3.0)",
    "email-validator (>=2.3.0,<3.0.0)",
    "authlib (>=1.6.5,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
